# docs/swagger.yaml
definitions:
  api.GetUserDetailOut:
    properties:
      age:
        type: integer
      balance:
        type: number
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  api.UserRecord:
    properties:
      age:
        type: integer
      balance:
        type: number
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  serializer.AddUserIn:
    properties:
      age:
        description: 年龄
        type: integer
      balance:
        description: 余额
        type: number
      name:
        description: 姓名
        type: string
    type: object
  serializer.UpdateUserIn:
    properties:
      age:
        description: 年龄
        type: integer
      balance:
        description: 余额
        type: number
      name:
        description: 姓名
        type: string
    type: object
  util.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/checkApikey:
    get:
      consumes:
      - application/json
      description: 检查Apikey
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - ApiKeyAuth: []
      summary: 检查Apikey
      tags:
      - 通用
  /api/v1/ping:
    get:
      consumes:
      - application/json
      description: Ping 测试接口
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
      summary: ping
      tags:
      - 通用
  /api/v1/user/add:
    post:
      consumes:
      - application/json
      description: 添加用户
      parameters:
      - description: 用户详情
        in: body
        name: string
        schema:
          $ref: '#/definitions/serializer.AddUserIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - ApiKeyAuth: []
      summary: 添加用户
      tags:
      - 用户管理
  /api/v1/user/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 删除用户
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除用户
      tags:
      - 用户管理
  /api/v1/user/detail:
    get:
      consumes:
      - application/json
      description: 获取用户详情
      parameters:
      - description: 用户id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.GetUserDetailOut'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取用户详情
      tags:
      - 用户管理
  /api/v1/user/list:
    get:
      consumes:
      - application/json
      description: 获取用户列表
      parameters:
      - description: 当前页数
        in: query
        name: current
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.UserRecord'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取用户列表
      tags:
      - 用户管理
  /api/v1/user/update/{id}:
    put:
      consumes:
      - application/json
      description: 更新用户信息
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户详情
        in: body
        name: string
        schema:
          $ref: '#/definitions/serializer.UpdateUserIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新用户信息
      tags:
      - 用户管理
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"